fix(commerce): urban_center_distance smallint → integer (overflow at >32km)
The commerce entity declared urban_center_distance as smallint while
lodgings/restaurants use integer. Owners entering distances ≥ 32768 meters
(≈32km) — common for commerces outside the urban core — hit
'value out of range for type smallint' on save.

Entity: changes the column decorator from 'smallint' to 'integer'.
Migration 1773900000000: ALTER COLUMN TYPE integer USING ::integer
— in-place catalog change, no row rewrite. Down migration refuses if
any row would lose precision (>32767), otherwise reverts to smallint.

Lodgings/restaurants already use integer — this brings commerce in line.

+++ src/modules/commerce/entities/commerce.entity.ts
@@ -122,7 +122,7 @@ export class Commerce {
   @Column('text', { name: 'google_maps_url', nullable: true })
   googleMapsUrl: string | null;
 
-  @Column('smallint', { name: 'urban_center_distance', nullable: true })
+  @Column('integer', { name: 'urban_center_distance', nullable: true })
   urbanCenterDistance: number | null;
 
   @Column('text', { name: 'how_to_get_there', nullable: true })
